#8 회원가입 페이지1 구현 - 수정사항반영 (스트링 추출, 텍스트필드 컴포넌트 등)
