+++ app/src/main/res/values/strings.xml
@@ -22,5 +22,9 @@
     <string name="btn_review_complete">완료</string>
     <string name="tv_review_length">%1$s/100자</string>
     <string name="toast_review_length">리뷰는 100자 이내로 작성해주세요.</string>
+    <string name="next">다음</string>
+    <string name="setIdPassword">아이디와 비밀번호를\n설정해주세요</string>
+    <string name="setId">아이디 입력</string>
+    <string name="setPassword">비밀번호 입력</string>
 
 </resources>
